// src/app/model/AddressModel.ts
export interface UserAddress {
    Address_ID?: string;         // Address ID
    User_ID: string;            // User ID
    Address_1: string;          // Address Line 1
    Address_2: string;          // Address Line 2
    District: string;           // District
    Province: string;           // Province
    Zip_Code: string;           // Zip Code
    Is_Default: boolean;
    Sub_District: string;       // Is Default
    Phone: string;
}